cns_project/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.10
django-address==0.2.8
django-bootstrap-v5==1.0.11
django-import-export==3.3.8
docopt==0.6.2
et-xmlfile==1.1.0
localflavor==1.9
MarkupPy==1.14
num2words==0.5.13
odfpy==1.4.1
openpyxl==3.1.2
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.10.0
xlrd==2.0.1
xlwt==1.3.0
